% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbwrapR-package.R
\docType{package}
\name{scbwrapR-package}
\alias{scbwrapR}
\alias{scbwrapR-package}
\title{scbwrapR: Wrapper for Statitics Sweden's API}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kira Coder Gylling \email{kira.gylling@gmail.com} (\href{https://orcid.org/0000-0002-9070-8381}{ORCID})

}
\keyword{internal}
